Move test cases from `TwigExtension` to `TwigRuntimeExtension`
